requirements.txt
ultralytics==8.1.25
streamlit
opencv-python
Pillow